Move doMultipleEdits() to Document
