+++ src/editor/Editor.js
@@ -1076,191 +1076,6 @@ define(function (require, exports, module) {
     };
 
     /**
-     * Adjusts a given position taking a given replaceRange-type edit into account. 
-     * If the position is within the edit range (start exclusive, end inclusive),
-     * it gets pushed to the end. Otherwise, if it's after the edit, it gets adjusted
-     * so it refers to the same character it did before the edit.
-     * @param {!{line:number, ch: number}} pos The position to adjust.
-     * @param {!Array.<string>} textLines The text of the change, split into an array of lines.
-     * @param {!{line: number, ch: number}} start The start of the edit.
-     * @param {!{line: number, ch: number}} end The end of the edit.
-     * @return {{line: number, ch: number}} The adjusted position.
-     */
-    Editor.prototype.adjustPosForChange = function (pos, textLines, start, end) {
-        // Same as CodeMirror.adjustForChange(), but that's a private function
-        // and Marijn would rather not expose it publicly.
-        var change = { text: textLines, from: start, to: end };
-
-        if (CodeMirror.cmpPos(pos, start) < 0) {
-            return pos;
-        }
-        if (CodeMirror.cmpPos(pos, end) <= 0) {
-            return CodeMirror.changeEnd(change);
-        }
-
-        var line = pos.line + change.text.length - (change.to.line - change.from.line) - 1,
-            ch = pos.ch;
-        if (pos.line === change.to.line) {
-            ch += CodeMirror.changeEnd(change).ch - change.to.ch;
-        }
-        return {line: line, ch: ch};
-    };
-    
-    /**
-     * Like _.each(), but if given a single item not in an array, acts as
-     * if it were an array containing just that item.
-     */
-    function oneOrEach(itemOrArr, cb) {
-        if (Array.isArray(itemOrArr)) {
-            _.each(itemOrArr, cb);
-        } else {
-            cb(itemOrArr, 0);
-        }
-    }
-    
-    /**
-     * Helper function for edit operations that operate on multiple selections. Takes an "edit list"
-     * that specifies a list of replaceRanges that should occur, but where all the positions are with
-     * respect to the document state before all the edits (i.e., you don't have to figure out how to fix
-     * up the selections after each sub-edit). Edits must be non-overlapping (in original-document terms).
-     * All the edits are done in a single batch.
-     *
-     * If your edits are structured in such a way that each individual edit would cause its associated
-     * selection to be properly updated, then all you need to specify are the edits themselves, and the
-     * selections will automatically be updated as the edits are performed. However, for some
-     * kinds of edits, you need to fix up the selection afterwards. In that case, you can specify one
-     * or more selections to be associated with each edit. Those selections are assumed to be in terms
-     * of the document state after the edit, *as if* that edit were the only one being performed (i.e.,
-     * you don't have to worry about adjusting for the effect of other edits). If you supply these selections,
-     * then this function will adjust them as necessary for the effects of other edits, and then return a
-     * flat list of all the selections, suitable for passing to `setSelections()`.
-     *
-     * @param {!Array.<{edit: {text: string, start:{line: number, ch: number}, end:?{line: number, ch: number}}
-     *                        | Array.<{text: string, start:{line: number, ch: number}, end:?{line: number, ch: number}}>,
-     *                  selection: ?{start:{line:number, ch:number}, end:{line:number, ch:number}, 
-     *                              primary:boolean, reversed: boolean, isBeforeEdit: boolean}>}
-     *                        | ?Array.<{start:{line:number, ch:number}, end:{line:number, ch:number}, 
-     *                                  primary:boolean, reversed: boolean, isBeforeEdit: boolean}>}>} edits
-     *     Specifies the list of edits to perform in a manner similar to CodeMirror's `replaceRange`. This array
-     *     will be mutated.
-     *
-     *     `edit` is the edit to perform:
-     *         `text` will replace the current contents of the range between `start` and `end`. 
-     *         If `end` is unspecified, the text is inserted at `start`.
-     *         `start` and `end` should be positions relative to the document *ignoring* all other edit descriptions
-     *         (i.e., as if you were only performing this one edit on the document).
-     *     If any of the edits overlap, an error will be thrown.
-     *
-     *     If `selection` is specified, it should be a selection associated with this edit.
-     *          If `isBeforeEdit` is set on the selection, the selection will be fixed up for this edit.
-     *          If not, it won't be fixed up for this edit, meaning it should be expressed in terms of
-     *          the document state after this individual edit is performed (ignoring any other edits).
-     *          Note that if you were planning on just specifying `isBeforeEdit` for every selection, you can
-     *          accomplish the same thing by simply not passing any selections and letting the editor update
-     *          the existing selections automatically.
-     *
-     *     Note that `edit` and `selection` can each be either an individual edit/selection, or a group of
-     *     edits/selections to apply in order. This can be useful if you need to perform multiple edits in a row
-     *     and then specify a resulting selection that shouldn't be fixed up for any of those edits (but should be
-     *     fixed up for edits related to other selections). It can also be useful if you have several selections
-     *     that should ignore the effects of a given edit because you've fixed them up already (this commonly happens
-     *     with line-oriented edits where multiple cursors on the same line should be ignored, but still tracked). 
-     *     Within an edit group, edit positions must be specified relative to previous edits within that group. Also,
-     *     the total bounds of edit groups must not overlap (e.g. edits in one group can't surround an edit from another group).
-     *
-     * @param {?string} origin An optional edit origin that's passed through to each replaceRange().
-     * @return {Array<{start:{line:number, ch:number}, end:{line:number, ch:number}, primary:boolean, reversed: boolean}>}
-     *     The list of passed selections adjusted for the performed edits, if any.
-     */
-    Editor.prototype.doMultipleEdits = function (edits, origin) {
-        var self = this;
-        
-        // Sort the edits backwards, so we don't have to adjust the edit positions as we go along
-        // (though we do have to adjust the selection positions).
-        edits.sort(function (editDesc1, editDesc2) {
-            var edit1 = (Array.isArray(editDesc1.edit) ? editDesc1.edit[0] : editDesc1.edit),
-                edit2 = (Array.isArray(editDesc2.edit) ? editDesc2.edit[0] : editDesc2.edit);
-            // Treat all no-op edits as if they should happen before all other edits (the order
-            // doesn't really matter, as long as they sort out of the way of the real edits).
-            if (!edit1) {
-                return -1;
-            } else if (!edit2) {
-                return 1;
-            } else {
-                return CodeMirror.cmpPos(edit2.start, edit1.start);
-            }
-        });
-        
-        // Pull out the selections, in the same order as the edits.
-        var result = _.cloneDeep(_.pluck(edits, "selection"));
-        
-        // Preflight the edits to specify "end" if unspecified and make sure they don't overlap. 
-        // (We don't want to do it during the actual edits, since we don't want to apply some of
-        // the edits before we find out.)
-        _.each(edits, function (editDesc, index) {
-            oneOrEach(editDesc.edit, function (edit) {
-                if (edit) {
-                    if (!edit.end) {
-                        edit.end = edit.start;
-                    }
-                    if (index > 0) {
-                        var prevEditGroup = edits[index - 1].edit;
-                        // The edits are in reverse order, so we want to make sure this edit ends
-                        // before any of the previous ones start.
-                        oneOrEach(prevEditGroup, function (prevEdit) {
-                            if (CodeMirror.cmpPos(edit.end, prevEdit.start) > 0) {
-                                throw new Error("Editor.doMultipleEdits(): Overlapping edits specified");
-                            }
-                        });
-                    }
-                }
-            });
-        });
-        
-        // Perform the edits.
-        this.document.batchOperation(function () {
-            _.each(edits, function (editDesc, index) {
-                // Perform this group of edits. The edit positions are guaranteed to be okay
-                // since all the previous edits we've done have been later in the document. However,
-                // we have to fix up any selections that overlap or come after the edit.
-                oneOrEach(editDesc.edit, function (edit) {
-                    if (edit) {
-                        self._codeMirror.replaceRange(edit.text, edit.start, edit.end, origin);
-
-                        // Fix up all the selections *except* the one(s) related to this edit list that
-                        // are not "before-edit" selections.
-                        var textLines = edit.text.split("\n");
-                        _.each(result, function (selections, selIndex) {
-                            if (selections) {
-                                oneOrEach(selections, function (sel) {
-                                    if (sel.isBeforeEdit || selIndex !== index) {
-                                        sel.start = self.adjustPosForChange(sel.start, textLines, edit.start, edit.end);
-                                        sel.end = self.adjustPosForChange(sel.end, textLines, edit.start, edit.end);
-                                    }
-                                });
-                            }
-                        });
-                    }
-                });
-            });
-        });
-        
-        result = _.chain(result)
-            .filter(function (item) {
-                return item !== undefined;
-            })
-            .flatten()
-            .sort(function (sel1, sel2) {
-                return CodeMirror.cmpPos(sel1.start, sel2.start);
-            })
-            .value();
-        _.each(result, function (item) {
-            delete item.isBeforeEdit;
-        });
-        return result;
-    };
-
-    /**
      * Sets the editors overwrite mode state. If null is passed, the state is toggled.
      *
      * @param {?boolean} start
